Manually fetch Localstack container image before running Kinesis specs
Previously, we relied on org.testcontainers' method of fetching container
image for Localstack. This was counted towards test timeout and rarely
actually triggered the tests.
Now, the image is fetched as part of beforeAll step and handled as part of
setup time that's not counted towards the timeout. If the image is already
available the cost of blocking is nearly zero.
